Support transport config proc in `Datadog::Profiling::Transport::HTTP`
This brings the `Datadog::Profiling::Transport::HTTP` behavior in line
with the `Datadog::Transport::HTTP` behavior: whenever a proc
is specified as `c.tracer.transport_options = proc { ... }` and can
be used by customers that need advanced configurations (such as
reporting to a unix socket).

In the future, it's expected that
`deprecated_for_removal_transport_configuration_proc` goes away and
the custom configuration is instead added as properties to the
`AgentSettings` object, but for now this is important to support so
that customers with an existing, working tracer configuration can
onboard to profiler without extra configuration.