Fix kwargs passing on Ruby 3.0 for sucker_punch
The kwargs separation change for Ruby 3.0 is tricky: there are different
behaviours on 2.6, 2.7, and 3.0. While ruby2_keywords attempts to smooth
that out, it is due to disappear on 3.2, therefore relying on it
produces surefire broken code when 3.2 is released and 2.6 EOL'd.

Thus the only future-proof and reliable syntax on Ruby >= 3 is to
capture the arguments with *args, **kwargs and forward them the same
way, without ruby2_keywords.

Conversely, the only working syntax on Ruby < 3 is capturing both with
*args and forwarding them the same way, using ruby2_keywords on 2.7 to
restore the previous behaviour.

Therefore either dynamic function definition per version is necessary
because of the syntactic difference in argument passing and forwarding.

An alternative is to cap the gem's maximum Ruby version.