Add benchmark for pprof encoding
While benchmarking discourse with profiler, I noticed a pattern of
increased latency on the app whenever a request happened concurrently
with pprof exporting.

A quick look with a profiler revealed that there seems to be some
low-hanging fruit in our exporting process, so I've added this
benchmark to allow me to iterate on improvements to this part of the
code.