Skip another test that needs fork on TruffleRuby
Although we don't officially support it, TruffleRuby seems to be able
to run profiler fine, including getting cpu time on Linux.

BUT the specs for CPU time profiling don't work, since they rely on
`fork()`, which TruffleRuby does not support. So for now let's skip
them, and in the future we can always revisit this if/when we decide to
officially support TruffleRuby.