Fix flaky cthread_spec tests due to races
There is a brief period where a thread has been started but the
`#native_thread_id` and `#clock_id` have not yet been set.
Because of it, a few of the tests in `cthread_spec.rb` could trigger
flaky failures from time to time.

This is fixed by ensuring that the thread has actually started and
set the needed ids before the expectations are run.

Link to failed CI run:
<https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/3505/workflows/2a54fe7e-8e43-4bd0-bd95-9be717eddf9f/jobs/138057/tests#failed-test-0>