Get rid of puts logging from profiler initialization
Before #1391, the `Datadog::Profiling::Tasks::Setup` task was run
before the gem logger was initialized/available, and thus the class
was made to use `puts` for its logging needs.

But since the refactor in #1391, this is no longer the case -- this
task is now called by
`Datadog::Configuration::Components#build_profiler` and thus we can
safely make use of the logger and get rid of any custom logging
specific to this class.