Tweak debug message about missing instrumentation
A customer was a bit misled by seeing this message in their logs
(e.g. it led them to believe CPU time profiling was not working).

This debug message is still useful for our own debugging so I've left
it, but I've reworded it to clarify that seeing this message is mostly
harmless.