Fix tracer ignoring value for service tag in DD_TAGS
As documented in
<https://docs.datadoghq.com/tracing/setup_overview/setup/ruby/#environment-and-tags>:

> If `DD_ENV`, `DD_SERVICE` or `DD_VERSION` are NOT set, tags defined in
> `DD_TAGS` will be used to populate `env`/`service`/`version` respectively.

Unfortunately, the previous implementation was missing a weird corner
case: the `service`/`env`/`version` in the settings object were only
set as a side-effect of trying to access or write to the tags option.

This meant that in a situation where `service`/`env`/`version` were
defined in `DD_TAGS` only AND those values were accessed BEFORE `tags`,
the accessing code would not see the correct values (it would still
see them as nil).

In particular, this happened during tracer initialization inside
`Components#build_tracer`, and meant that rather than using the service
name defined in `DD_TAGS`, the fallback default from
`Tracer#default_service` would be used instead.

I'm not very happy with the fix (hence it's more of a workaround),
see the huge comment I left on the code for more details on why
I'm not happy and why it works.