Explicitly start profiler on ddtrace/auto_instrument
For users that enable profiling via environment variable
(`DD_PROFILING_ENABLED=true`), doing
`require 'ddtrace/auto_instrument'` means that the profiler is also
started as a side effect, even if they forget to use `ddtracerb exec`
or `require 'ddtrace/profiling/preload'`.

This is a nice property!

Right now it semi-happens by chance, but I want to make it an actual
feature, so I'm adding a bit more code and tests to make this happen
explicitly so that we don't regress.