[PROF-3514] Skip CPU time instrumentation if logging gem is detected
Instead, customers will get the following log message:

```
INFO -- ddtrace: [ddtrace] (/app/lib/ddtrace/profiling/tasks/setup.rb
:44:in `activate_cpu_extensions') CPU time profiling skipped because
native CPU time is not supported: The `logging` gem is installed and
its thread inherit context feature is enabled. Please add
LOGGING_INHERIT_CONTEXT=false to your application's environment
variables to disable the conflicting `logging` gem feature. See
https://github.com/TwP/logging/pull/230 for details. Profiles
containing Wall time will still be reported.
```

This is a repeat of #1362 -- the logging gem also monkey patches
`Thread` initialization using `alias_method`, which is incompatible
with our instrumentation that uses `preload`.

For details, see https://github.com/TwP/logging/pull/230 where we
submitted a fix to the upstream gem.

Right now the only way to avoid this issue and still enable proflier is
to either remove the logging gem or to disable the feature using the
`LOGGING_INHERIT_CONTEXT` environment variable.

If/when the gem creators accept our proposed change and release it,
we'll need to revise this to take that into consideration, as we do
for rollbar.

NOTE: I'm not entirely happy with the complexity that is accumulating
for the tests, but I've decided to not refactor it yet. If we need
to add more gems, we should definitely refactor the tests. My hope is
that soon we can get rid of our `Thread` monkey patches, and thus get
rid of all of this instead.