Require agent_settings argument in Profiling::Transport::HTTP
The `agent_settings` is allowed to be `nil` whenever the agentless
configuration is used, but let's make it clearer that the
default/expected path is not to use agentless, and force callers to
pass in `agent_settings: nil`, rather than using that as a default.

I've also refactored the tests to reflect that this is the primary
usecase.