[PROF-2720] Remove automatic agentless support
The profiler supports reporting profiling data directly to the
datadog backend, without having to go through an agent. This setup is
actively discouraged by the profiling team for users, but hey, it is
really handy for testing and prototyping.

But we don't want users to accidentally enable this mode. Up until now,
profiler would implicitly use this mode if users set a `DD_API_KEY` and
a `DD_SITE` configuration. Now, this mode will only be enabled if
a `DD_PROFILING_AGENTLESS` environment variable is set to `true`.