Upgrade outdated datadog agent used in CI testing
Oops! We were actually using a really legacy agent for CI testing,
which did not match the version in docker-compose.yml.

This explains why a profiler integration test was failing in CircleCI
but not in local testing: in CircleCI we were using an agent version
so old that it did not have profiling support!