Add experimental toggle to disable the profiling native extension
In #1584 I identified that users of dd-trace-rb on Linux/macOS
needed to have a working build environment to compile msgpack and ffi,
and so installing our gem would be no different.

I've since identified a possible corner case: in the reliability
environment, we use the GitLab omnibus distribution
(<https://docs.gitlab.com/omnibus/>), which already ships
pre-compiled versions of msgpack and ffi, but does not include a
compiler nor the Ruby headers required for compiling extensions.

Thus, up until now it was trivial to add dd-trace-rb (and profiling)
to these images, but the addition of the native extension complicates
things a lot.

Since the native extension is still in the "we're evaluating and
learning about the downsides/sharp edges of having it" phase, I
decided to add a `DD_PROFILING_NO_EXTENSION` environment variable,
that for now can be used to disable building and usage of the
extension.

Because the extension is currently empty, doing this is is harmless,
but in the future, the use of `DD_PROFILING_NO_EXTENSION` will lead
to the profiler being disabled as well, once the profiler code
starts requiring the native extension to be available. For that
reason, I documented this toggle as experimental, and added a
warning whenever the app is loaded with it.

For users reading this (if any!): if you needed to use this option,
please tell us why on
<https://github.com/DataDog/dd-trace-rb/issues/new> as we want to
acommodate your use cases.