Extract resource name from ddtrace HTTP::TYPE_INBOUND traces
This extends the previous behavior of TraceIdentifiers::Ddtrace to also
return the root span resource, which will be used to enable
aggregation-by-endpoint functionality in the profiler.

Due to PII concerns (as explained in the comments), we start with a
rather conservative filter of only allowing HTTP::TYPE_INBOUND traces.

I took some inspiration from the similar implementation for Python
(<https://github.com/DataDog/dd-trace-py/pull/2573>) although I still
need to validate with the wider team that the current filtering
approach is reasonable for Ruby tracing as well.