RFC: Add Sorbet typechecker to dd-trace-rb
 ## TL;DR

* Minimal bootstrapping for now.
* Type checking is optional. Default is no type checking for files
  without a magic comment. Should have minor impact on people that
  don't want to touch the type checking.
* `bundle exec srb tc` runs the type checking.
* I've enabled type checking during CI runs

 ## What?

While not a replacement for all our tests and other tooling, a type
checker is a nice extra line of defense for catching issues that we
otherwise may have missed, such as #1602.

The Ruby ecosystem around type checking is still rather immature. You
may have heard that Ruby 3 shipped "with type checking" but that's an
oversimplification: it actually only shipped with a "language for
describing types" and a "type analysis tool"
(<https://www.ruby-lang.org/en/news/2020/12/25/ruby-3-0-0-released/>).
You may have noticed that nowhere in the previous sentence the words
"type checker" were included.

So, Ruby 3 itself provides no type checker. In practice, there are two
typecheckers in "wide" use for Ruby:

* [Steep](https://github.com/soutaro/steep), created by a Ruby core team
  member that also worked on manny of the other type-related bits that
  made it to Ruby 3 (but steep itself was not made part of Ruby)
* [Sorbet](https://sorbet.org/), a type checker in wide use at Stripe
  and Shopify

This RFC PR does a first pass at onboarding Sorbet to our codebase.
There's a number of limitations on this "first pass", which I document
below.

 ## Why not Steep?

I've tried onboarding our codebase to steep, and gave up due to two big
issues:

1. Steep relies on the Ruby 3 rbs format, but unlike Sorbet, there
doesn't seem to be any way of starting small on a mostly-untyped
codebase without any `.rbs` files, and iterate from there. There are
tools to generate skeleton `.rbs` files (the `rbs` gem has two
different modes for it + the `typeprof` gem), but they had a hard time
with ddtrace.

2. There is currently little documentation for the type checking errors,
once Steep does run. Google gave me ZERO hits on one of them (I guess
it doesn't index the tool repository at least). And they're not all
that easy to understand/fix, so what use is having a tool tell
you "this is wrong", if then you're stuck with little help on something
that may not even be a real issue but more likely a tool limitation?

For these reasons, I gave up on Steep (for this first iteration).

 ## Working with Sorbet

Sorbet can be run with `bundle exec srb tc` (or `bundle exec rake
typecheck`). There's also Language Server Protocol support, if your
editor supports it.

The key thing to know about Sorbet is that allows file-level granularity
for enabling type checks, see
<https://sorbet.org/docs/static#file-level-granularity-strictness-levels>.

In practice, this means that at the top of each `.rb` file, we can
provide a comment that controls the reporting of type checking errors
for that file. The usual ones (described in the above link as well) are:
`# typed: ignore`, `# typed: false`, `# typed: true`, `# typed: strict`,
and `# typed: strong`.

In many cases, **Sorbet can typecheck a file correctly with no extra
type annotations**, so we can start enjoying its checks (such as finding
 #1602) by just enabling it for our files. This can be done with the `#
 typed: true`, `# typed: strict`, `# typed: strong` modes (see the docs
 for details).

The `# typed: false` is the default for files with no comment. Thus,
even newcomers into the codebase will not be forced to immediately deal
with Sorbet for contributions, which given the current limitations of
the tool, and level of maturity, I think is reasonable. (We can
reconfigure this later, if needed). In this mode, Sorbet will only
complain about syntax and missing constant errors.

Our codebase requires quite a few `# typed: ignore` because or
dependencies that are pulled via appraisals and thus not seen/available
to Sorbet (among other things that we do and that Sorbet) doesn't quite
like. These files are completely ignored by Sorbet. (Note that if you
then reference something on one of these files from another file, that
file will not typecheck either, so "ignore" becomes somewhat contagious,
and is best avoided if possible).

I've also added a CI step to run sorbet, which again follows the rules
above so should be able to provide us with valuable insight without
being an annoying source of build failures.

 ## Limitations or "I thought the whole point of this was for us to
   add type declarations to the code"?

Now we get to the more unfortunate part of the news: I think that for
now we don't have great options for doing this.

Sorbet type annotations are supposed to be inline with the code, and
work through the [`sorbet-runtime`]
(https://rubygems.org/gems/sorbet-runtime) gem.

Unfortunately, as described on
<https://sorbet.org/docs/faq#what-platforms-does-sorbet-support>:
> The sorbet-runtime gem is currently only tested on Ruby 2.6 and Ruby
  2.7. It is known to not support Ruby 2.4. Feel free to report runtime
  issues for any current or future Ruby version.

Thus, since we still want to support Ruby 2.1, that doesn't quite work
for us. I thought we could perhaps have an empty replacement of the
sorbet runtime gem, and it turns out that [Shopify already tried and
abandoned it](https://github.com/Shopify/sorbet-runtime-stub). So,
inline annotations seem to be out for us.

For Sorbet, that leaves us with only one other option: `.rbi` files.
These are files that allow Sorbet to get type information for otherwise
untyped code, and are the solution to having type information for many
common libraries.

Unfortunately, using `.rbi` files for our own code is very error-prone,
because Sorbet trusts `.rbi` files, and doesn't check them against the
code. Thus, if we had a class `A` with method `m` declared both on our
code, as well as on an `.rbi` file, and then we renamed `m` to `n` and
forgot to update the `.rbi`, then Sorbet would not complain about it.
Worse, if somewhere there was a call to `A#m` somewhere in the code,
Sorbet would still think it was correct, while in reality it would
break at run-time.

 ## The future?

Because our usage of Sorbet is rather simple, we can at any point
switch to Steep, or any other tool, or just disable it, if it's not
providing value. So we are not locked at all into its usage.

Hopefully in the near future we'll find a way to add type declarations
to our library without impact to our support matrix of older Rubies.

Even better, we could then ship those types to our customers, for them
to check that they are using ddtrace APIs correctly!