Run `bundle exec srb init`
Everything on this commit was 100% generated by Sorbet.

Sorbet checked all our files and added the best `# typed: ...`
comments it could for each (e.g. if it could type it, it added
`true`, if it couldn't, it added `false`, etc).

Included in here is also the `sorbet/` folder which is used
to keep type information for our dependencies.