Add sidekiq to rack CI test app
This commit adds a very simple setup that allows us to write some key
and value to redis asynchronously (using a sidekiq background job)
and to read it synchronously.

This is used to drive a very simple test via two new endpoints,
allowing us to trigger a new sidekiq task (via the write_sidekiq
endpoint), and to observe its successful conclusion (via the
read_sidekiq endpoint).