Add resque to rack CI test app
Mirroring our sidekiq test, this commit adds a very simple setup
that allows us to round trip some information through a resque
background worker, and then return it to be asserted in our tests.

This is used to drive a test that resque is working correctly with
dd-trace-rb, and that profiling is enabled and working in the resque
process.