Store resource in ResourceContainer object
Some integrations only set the resource name at the end of the request.
When the profiler is taking a sample of a given thread with an active
span, the resource name may not yet be available, or it may be
incorrect.

By introducing the `ResourceContainer` object in a span, the profiler
can keep a reference to the `ResourceContainer` objects of all root
spans it observes during sampling, and thus delay getting the
resource name to as late as possible.

Hopefully by the time the profiler is preparing a submission to the
backend, the final correct resource name is available, and it can
be used by the profiler.

Note that this still doesn't guarantee 100% correctness, since the
profiler will not delay sending a profile while a span is still
ongoing, but this approach should provide us with a reasonable
approximation.