Set Sinatra resource setting at beginning of request
As of #1623, the profiler records the `resource` of the root span,
on top of the span id and trace id during sampling.

Also, as discusssed in the description of #1623 (in the
"Getting the correct `resource`" section) integrations where the
`resource` is only set at the end pose an extra challenge, as the
request may not be finished in time for the `resource` to be included
in the profiler payload.
This means that the profiler may miss the `resource` for some of the
requests, depending on timing of when the sampling happens and when the
request finishes.
(Note that in this case, the trace id and span id will still be
propagated, so the samples will always be able to be tied back to the
trace that originated them; it's just the `resource` that will not be
included in the profiler data).

To avoid this issue for many of our customers, let's set the Sinatra
`resource` setting to the beginning of the request, thus avoiding
the issue altogether for Sinatra.

Note that I did not touch the existing code to set the resource at the
end, which had a quite scary TODO (lines 47-64) and seemed like
something I shouldn't touch.

Note 2: This is similar to what we did in #1626 for the Rails
integration.