[PROF-3314] Do not report profiles with little data (duration < 1s)
Due to the way that ddtrace configuration works, if profiling is
started using DD_PROFILING_ENABLED=true but the app is then
reconfigured using Datadog.configure , the profiler tries to report an
almost empty profile (with < 1s of data usually).

This can be even worse if the Datadog.configure call was also setting
up the configuration for reporting to the agent, because this empty
profile report will be using the environment defaults and thus result
in an error if they are incorrect.

Since these profiles are of little value, let's just skip reporting
them.