Simplify cleaning of leftover cpu time tracking state
Since this method is being called at stack sampler startup, threads
that previously were not touched were going to get written to anyway
on the next sample.

The one exception/behavior change is if cpu time profiling is
disabled or unavalable for some reason -- we would previously never set
the `THREAD_LAST_CPU_TIME_KEY` for threads in that case, but I don't
think it's worth having this distinction, so I'll take the simpler
implementation and tests instead.