Add spec to validate that benchmarks don't break
While working on #1648 I discovered that one of the profiler benchmarks
(profiler_submission.rb) was actually broken, and had been for some
time.

To avoid these "surprises", I've added a small spec to load and run
the benchmarks, so that if they start failing, we'll know about it.

To avoid impacting the speediness of our test suite, the benchmarks
are run with a minimal time/warmup -- the objective is not to measure
anything at all, but just to confirm that they do run.

P.s.: This PR was opened on top of #1648; once that branch is merged,
this one will switch to target master as well.