Do not skip profiler tests if profiling does not seem to be working
Profiling seeming not to be working is an important symptom --- in our
development setup and in CI profiling is always supposed to be working
(unless we're on JRuby), so we should not just smile and wave when
something is going wrong.

See also the previous commit for an example of something that could
go wrong and silently make all tests not execute, but still pass.