Fix CI sorbet failure and lock down sorbet version
Sorbet releases almost daily, and thus when I added it, I set the
acceptable versions to '>= 0.5.6513', '< 0.6' and let new
releases in, as I assumed that they may bump the version if
incompatible changes landed.

Twice so far this hasn't been the case, so let's switch gears: we now
lock to a specific version, and we'll need to bump it from time to
time to avoid falling too far behind.