Auto-build profiling native extension before running integration apps locally
In #1615 I added profiler tests to the integration test apps, and made
sure that the profiling native extension was built when running the ci
scripts.

I completely missed that the extension was not built when running the
integration apps locally.

I've now extracted the "build the native extension" into a separate
script in the shared folder; and made sure that it runs for both CI
and also when running the apps locally using docker.