[PROF-3944] Change profiler to gather "local root span id" from active traces instead of "trace id"
The `trace id` that we gather today is not useful, because the
backend ends up using the `runtime-id` and `span id` to get all
information it needs.

Gathering instead the `local root span id` will enable the profiler
backend to optimize scoping profiles in the UX down to a single
request (all spans that share the same local root span id).