Update profiler-submission benchmark with dump that uses root span id
Otherwise the previous benchmark dump was innacurate, since the
resulting pprof no longer cared for the trace id that it included.

I've renamed the benchmark to "profiler_submission_v2" to mark the
change to the input data -- I've measured the performance of the
benchmark to stay similar, but not the same [within ~10%], so it's
relevant to mark this change.