Track last wall time per-thread via thread-local variables
Previously, we kept the `last_wall_time` as an instance variable of
the Stack collector, and on every call to `#collecet_events` we would
gather the current wall time, subtract from it the last wall time, and
use that as the `wall_time_interval_ns` for all sampled threads.

As preparation for being able to not sample all threads at once, we
need to start storing the last wall time per-thread, so that if
a thread is not chosen on a call to `#collect_events` to be sampled,
but is chosen on the next one, the resulting wall clock count is
still accurate.

E.g. consider if a thread is sleeping, and we sample once per second
BUT only sample that thread every other second. Without this change,
the thread would be assigned 30s of wall time (incorrect); with this
change, it is correctly assigned 60s of wall time.

Note that effectively we already did this for cpu time calculation,
and hence the refactoring basically just changed the wall time to
use the common `#get_elapsed_since_last_sample_and_set_value` helper
function.