Fix flaky SuckerPunch patcher spec on JRuby
This took me a while to hunt down, and it manifested as a
`StandardError: Wait time exhausted!` in the
`try_wait_until { Thread.list.size < count }` line of the `after`
cleanup block.

Example failure:
<https://app.circleci.com/pipelines/github/DataDog/dd-trace-rb/4346/workflows/88a101cc-8c06-44b6-84b6-efedef4e426a/jobs/165493/parallel-runs/1>

The issue was caused by the cleanup code assuming that the Thread list
would go down after telling SuckerPunch to shut down.

This assumption was broken by test cases where we run the test so fast
that SuckerPunch doesn't get to start a background thread to run the task.
Thus, the number of threads in the `after` block can never go down,
since none ended up being started.

One way of simulating this issue is to comment out the
`__perform_async(*args)` in the `Instrumentation` file, so that the
enqueue span gets created, but no perform ever gets called.

The fix is to ensure that the SuckerPunch task actually gets executed.