[PROF-4081] Make threads with empty backtrace visible
As documented in the code, there's a few situations in which a thread
can have an empty backtrace. Previously, these threads were not visible
in the profiling UX, which was a gap in our provided information.

(It also could be confusing, for instance, if a thread spent some time
executing Ruby and some time in native code, as happens with the Iodine
web server, the "wall clock time" would not add up to 60s in a regular
profile, because we ignored these threads).

To fix this, we now replace empty backtraces with a fake stack that
displays `(In native code)`. We currently have no way of showing
what's up in native code (is the code running? is the thread idle?
we don't know), but at least we know it's there.