Separate out building native extension from tests on Windows
It looks like we have a number of breakages of our test suite on
Windows, unrelated to this PR (this is probably the first time our
test suite is being run on Windows...).

To avoid blocking the current task, I'll enable validation of the
native extension build on Windows, and make that break CI if it fails;
but will allow failures in the regular test suite until we can fix them.