Add deprecation warning for Ruby 2.1 support
This PR, modelled on #1441, adds a deprecation warning for Ruby 2.1
users of ddtrace.

Ruby 2.1 has not received security updates since the
[1st of April, 2017](https://www.ruby-lang.org/en/news/2017/04/01/support-of-ruby-2-1-has-ended/)
and we, like the core team, highly recommend upgrading to a later
version.

Support for Ruby 2.1 will be dropped in a future release, probably
`0.54.0`.

Users are welcome to continue to use `< 1.0.0` for their Ruby 2.1
deployments going forward.