Break CI run if integration test fails
Since #1526 we've been printing the app logs when a failure is
triggered, but unfortunately the behavior of using

```some failure || print logs```

breaks the bash `set -euo pipefail` behavior we use that should cause
the `ci` test scripts to fail if the underlying command fails.

I've noticed this because there are a few CI integration tests that
are actually broken and CI is still showing up as green.

By adding the `exit -1`, the intended failure behavior is
restored.