Refactor ci task in Rakefile for better readibility
I've been thinking about this for a long time, and today I had a bit of
time between meetings to do it.

This PR squashes the per-Ruby appraisal declarations and replaces them
with a single string of Rubies to be enabled/disabled, thus massively
simplifying the Rakefile.

The actual list of tasks was generated using code: I copy pasted
the previous declaration, and implemented a few alternative methods
that kept track of what was enabled per-Ruby. In a second pass,
I manually reordered a few items, restored the previous comments, and
marked the items that are not usable on JRuby.

Thus, I'm quite confident that the output is equivalent to what
we had before.