Skip building profiling native extension when Ruby has been compiled without JIT
The profiling native extension relies on the Ruby MJIT headers on 2.6+.

But these headers are not available in Rubies that were compiled with
JIT disabled.

To avoid impacting these users (e.g. making them explicitly need
to install with `DD_PROFILING_NO_EXTENSION=true`), let's
automatically skip the extension when we detect that the JIT was
disabled.

In the future, we can explore using `debase-ruby_core_source` for
these Ruby versions, if we still want to support these users.

Fixes #1774