Workaround missing dependency on mail gem in Ruby 3.1
The `net-smtp` gem used to be built-into Ruby, but was extracted in
Ruby 3.1. The `mail` gem (which is a dependency of a bunch of Rails
stuff) relies on this gem, but did not explicitly depend on it, and
so it broke.

There's an upstream PR to fix this:
<https://github.com/mikel/mail/pull/1439> but while it's not
released, let's just pull in `net-smtp` directly.