Send number of allocations in span as metric
I recently noticed that while we usually collect the number of memory
allocations on a given span, we send it as an "allocations" field to
the agent and this data never seems to make it to the backend.

Instead, this refactor reports it as a span metric, which makes
it available to customers.