[PROF-4260] Add jitter to profile flushes
In Ruby applications that use multiple processes (such as webservers
like puma), we start a profiler instance for each individual process.

If those applications fork all their processes at the same time
(usually at start-up time), we get N processes that start profiling
at the same time, and that will report profiles every minute
at exacty the same time.

To spread out the impact of this reporting (both on the reporting app,
as well as on the profiling backend), this PR adds a very small
random sleep (at most 3 seconds) before each report is made.

Note that because we sleep AFTER we collect the events from the
recorder, we still report exactly the same data as before -- we just
may report it ever-so-slightly later.