Suite as example group file path (shared_examples)
In Datadog we report all tests' suite names as the example file path,
which usually is the same as the example group file path. This is not
always true however, especifically on the case of tests included from a
different path (usually on a shared_examples block).

Because of that, suite name would be reported as the shared examples
file path, which was not not very helpful to track the files that are
taking the most to run as the shared examples can be on a different
path than the test file that included them.

To address this, the this commit adds the use of
`metadata[:example_group][:file_path]`, so all test suites should be
reported as the _example group file path_.