Fix flaky profiling/validate_benchmarks spec
The intention of the `profiling/validate_benchmarks` spec is to run the
profiler benchmarks for a short amount of time as part of our test
suite, to make sure they haven't broken.

But we saw some flakiness in these tests (see #1812) due to the
benchmark harness failing with

```
/usr/local/bundle/gems/benchmark-ips-2.9.1/lib/benchmark/ips/job.rb:187:in `to_i': Infinity (FloatDomainError)
```

Upon investigation, this was caused by the warm up time being so
small that sometimes it did not run, which benchmark-ips didn't like
at all -- see https://github.com/evanphx/benchmark-ips/issues/120 .

To fix this, I've disabled the warmup (set it to 0), as we don't
really care about warmup on these tests.

I've also noted (in <https://github.com/evanphx/benchmark-ips/issues/120#issuecomment-1006487480>)
that benchmark-ips also doesn't like if the test time is too low,
so I've bumped it a bit up so that hopefully we'll never trigger
that issue either.

Fixes #1812